+++ plugins/lint/src/interfaces.ts
@@ -1,4 +1,6 @@
-export interface LintArgs {
+import stylelint from 'stylelint';
+
+interface LintArgs {
     /** Attempt to fix the lint errors. */
     fix?: boolean;
     /** Do not use any cached results from previous runs. */
@@ -7,4 +9,24 @@ export interface LintArgs {
     annotate?: boolean;
     /** An optional list of files to lint */
     files?: string[];
-  }
+}
+
+type StylelintResult = stylelint.LinterResult & {
+    /** Lines with needless disables */
+    needlessDisables?: {
+        /** The files the needless disables are found in */
+        source: string;
+        /** Ranges with needless disables */
+        ranges: {
+            /** Rule name */
+            unusedRule: string;
+            /** Line number */
+            start: number;
+        }[];
+    }[];
+};
+
+export {
+    LintArgs,
+    StylelintResult
+}
