+++ plugins/lint/src/utils/helperUtils.ts
@@ -1,84 +1,64 @@
 import getPackages from 'get-monorepo-packages';
 import fs from 'fs';
-import stylelint from 'stylelint';
 import path from 'path';
-
-
 import {
-    getMonorepoRoot,
-    createLogger  } from '@design-systems/cli-utils';
-import {LintArgs} from '../interfaces';
-
-
-  type StylelintResult = stylelint.LinterResult & {
-    /** Lines with needless disables */
-    needlessDisables?: {
-      /** The files the needless disables are found in */
-      source: string;
-      /** Ranges with needless disables */
-      ranges: {
-        /** Rule name */
-        unusedRule: string;
-        /** Line number */
-        start: number;
-      }[];
-    }[];
-  };
+  getMonorepoRoot
+} from '@design-systems/cli-utils';
+import { LintArgs, StylelintResult } from '../interfaces';
+import stylelint from 'stylelint';
 
 /** Get the folders that contain packages in the monorepo. */
 function getPackageFolders() {
-    return [
-      ...getPackages('.').reduce((all, p) => {
-        all.add(p.location.split('/')[0]);
-        return all;
-      }, new Set<string>())
-    ];
-  }
+  return [
+    ...getPackages('.').reduce((all, p) => {
+      all.add(p.location.split('/')[0]);
+      return all;
+    }, new Set<string>())
+  ];
+}
 
-  /** Determine when a file was last edited */
+/** Determine when a file was last edited */
 function getLastEdited(filename: string) {
-    try {
-      return fs.statSync(filename).mtimeMs;
-    } catch (error) {
-      return 0;
-    }
+  try {
+    return fs.statSync(filename).mtimeMs;
+  } catch (error) {
+    return 0;
   }
+}
 
-  /** Return a list of files to lint */
+/** Return a list of files to lint */
 function files(args: LintArgs, types: string) {
-    const isRoot = getMonorepoRoot() === process.cwd();
-  
-    if (args.files) {
-      const fileExt = new Set(types.split('|'));
-      return args.files.filter(f => fileExt.has(path.extname(f).substr(1)));
-    }
-  
-    if (isRoot) {
-      return [`components/**/src/**/*.(${types})`];
-    }
-  
-    return [`src/**/*.(${types})`];
+  const isRoot = getMonorepoRoot() === process.cwd();
+
+  if (args.files) {
+    const fileExt = new Set(types.split('|'));
+    return args.files.filter(f => fileExt.has(path.extname(f).substr(1)));
   }
 
-  /** Run stylelint with the given options. */
+  if (isRoot) {
+    return [`components/**/src/**/*.(${types})`];
+  }
+
+  return [`src/**/*.(${types})`];
+}
+
+/** Run stylelint with the given options. */
 async function attemptStylelint(
-    options: Partial<stylelint.LinterOptions>
-  ): Promise<StylelintResult> {
-    try {
-      return (await stylelint.lint(options)) as StylelintResult;
-    } catch (error) {
-      if (error.message.includes('No files matching the pattern')) {
-        return {
-          results: [],
-          output: 'no match',
-          errored: false
-        };
-      }
-  
-      throw error;
+  options: Partial<stylelint.LinterOptions>
+): Promise<StylelintResult> {
+  try {
+    return (await stylelint.lint(options)) as StylelintResult;
+  } catch (error) {
+    if (error.message.includes('No files matching the pattern')) {
+      return {
+        results: [],
+        output: 'no match',
+        errored: false
+      };
     }
+
+    throw error;
   }
+}
 
-
-
-export {getPackageFolders, getLastEdited, files, attemptStylelint, getMonorepoRoot};
+export { getPackageFolders, getLastEdited, files, attemptStylelint, getMonorepoRoot };
